Implemented KHR_texture_transform extension. (#24)
* Implemented KHR_texture_transform for primitive fragment shader (unlit·depth prepass·jump flood seed shader not yet).

* New function: fastgltf::getTexcoordIndex(get texture coordinate index from fastgltf::TextureInfo with respecting KHR_texture_transform)

* Added GUI for transformed texture coordinate information.

* Added reflect module, adopted it to AggregateHasher for automatic struct field count use.

* Use static reflection based vk::SpecializationInfo generation and re-structured the shader variants.

* Implemented unlit material texture transform.

* Finished KHR_texture_transform implementation by applying it to MaskDepthRenderer and MaskJumpFloodSeedRenderer.

* MSVC workaround

* Fix lifetime issue.

* Fix vertex color regression.

---------

+++ shaders/unlit_primitive.vert
@@ -15,7 +15,7 @@
 #define HAS_VARIADIC_OUT HAS_BASE_COLOR_TEXTURE || HAS_COLOR_ATTRIBUTE
 
 layout (constant_id = 0) const uint TEXCOORD_COMPONENT_TYPE = 5126; // FLOAT
-layout (constant_id = 1) const uint8_t COLOR_COMPONENT_COUNT = uint8_t(0);
+layout (constant_id = 1) const uint COLOR_COMPONENT_COUNT = 0;
 layout (constant_id = 2) const uint COLOR_COMPONENT_TYPE = 5126; // FLOAT
 
 layout (std430, buffer_reference, buffer_reference_align = 1) readonly buffer U8Vec2Ref { u8vec2 data; };
